Separate front and bottom nodes of the gate task.
