+++ vision_layer/vision_tasks/cfg/gateBottomRange.cfg
@@ -1,0 +1,22 @@
+#!/usr/bin/env python
+PACKAGE="vision_tasks"
+
+from dynamic_reconfigure.parameter_generator_catkin import *
+
+gen = ParameterGenerator()
+
+gen.add("clahe_clip", double_t, 0, "CLAHE clip limit", 4.0, 0.0, 40.0);
+gen.add("clahe_grid_size", int_t, 0, "CLAHE grid size iterations", 8, 1, 16);
+gen.add("clahe_bilateral_iter", int_t, 0, "Bilateral filter on CLAHE iterations", 8, 0, 16);
+gen.add("balanced_bilateral_iter", int_t, 0, "Bilateral filter on white balanced iterations", 4, 0, 8);
+gen.add("denoise_h", double_t, 0, "Denoising h value", 10.0, 0, 20.0);
+gen.add("low_b", int_t, 0, "Lower Bound Of B", 254, 0, 255);
+gen.add("high_b", int_t, 0, "Higher Bound Of B", 255, 0, 255);
+gen.add("low_g", int_t, 0, "Lower Bound Of G", 254, 0, 255);
+gen.add("high_g", int_t, 0, "Higher Bound Of G", 255, 0, 255);
+gen.add("low_r", int_t, 0, "Lower Bound Of R", 254, 0, 255);
+gen.add("high_r", int_t, 0, "Higher Bound Of R", 255, 0, 255);
+gen.add("closing_mat_point", int_t, 0, "", 1, 1, 7);
+gen.add("closing_iter", int_t, 0, "Closing iterations", 0, 0, 10);
+
+exit(gen.generate(PACKAGE, "vision_tasks", "gateBottomRange"));
